修改用户登录/注册接口

+++ src/main/java/com/centrin/ciyun/medrpt/param/PersonBaseInfoParam.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class PersonBaseInfoParam extends BaseEntity{
 	private String nickname;//用户昵称
-	private int sex;//用户性别 1：男 2：女 3：未知
-	private int age;//用户年龄
-	private int hight;//用户身高
+	private Integer sex;//用户性别 1：男 2：女 3：未知
+	private Integer age;//用户年龄
+	private Float hight;//用户身高
 }
